delete scene method will raise bad request when scene id is not provided

+++ svc/db/methods/user_credentials.py
@@ -235,6 +235,8 @@ class UserDatabase:
 
     def delete_scene_by_user(self, user_id, scene_id):
         self.__validate_user_id(user_id)
+        if scene_id is None:
+            raise BadRequest()
         self.session.query(SceneDetails).filter_by(scene_id=scene_id)
         self.session.query(Scenes).filter_by(user_id=user_id, id=scene_id).delete()
